root/requirements.txt
fastapi>=0.100,<1.0
uvicorn[standard]>=0.23
pydantic>=2.6,<3
pydantic-settings>=2,<3
sqlalchemy[asyncio]>=2.0,<3
asyncpg>=0.29
alembic>=1.13
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
email-validator>=2.0
httpx>=0.27
python-multipart>=0.0.7
aiofiles>=23.0